scripts/requirements.txt
icalendar
beautifulsoup4
requests
fire
pyyaml